Add files via upload
Fixed the issue with the map not full y loading, added the highscore text, created functions to make the code in map more organized. Added a few comments in javaaDoc format.

+++ Grid.java
@@ -25,13 +25,13 @@ public class Grid {
 	
 	public void addCell(Cell cell){ //we need to make copies of the cell because encapsulation
 		
-		grid [cell.getCoordinate().getXlocation()][cell.getCoordinate().getYlocation()] = cell;
+		grid [(int)cell.getCoordinate().getXlocation()][(int)cell.getCoordinate().getYlocation()] = cell;
 		
 	}
 	
-	public  Cell getCell(int row, int column){
+	public  Cell getCell(double row, double column){
 		
-		return grid [row][column];
+		return grid [(int)row][(int)column];
 		
 	}
